LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.2_White_Space/S7.2_A5_T5-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unrecognized token '\u00A0'
S7.2_A5_T5

PASS Expected parsing failure

TEST COMPLETE
